# feat: bounded concurrency for binary inspection