@Gelio 2023-09-13
Hey, thanks for the contribution! Sorry for lack of activity on this PR. I am on an extended vacation.

The code looks good at first sight. I will do a more thorough review of it next week.

I see the CI failed for go 1.17. I presume it is a problem with [shfmt](https://github.com/mvdan/sh) being incompatible with go 1.17, and not anything related to this change. You do not need to worry about it