@Gelio 2023-09-19
I have published this improvement in [v0.2.3](https://github.com/Gelio/go-global-update/releases/tag/v0.2.3)